requirements.txt
Django==1.8
SQLAlchemy==1.0.2
dj-database-url==0.3.0
dj-static==0.0.6
django-extensions==1.5.2
django-postgrespool==0.3.0
djangorestframework==3.1.1
gunicorn==19.3.0
psycopg2==2.6
pydot
pyparsing==1.5.7
six==1.9.0
static3==0.5.1
wsgiref==0.1.2
